% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_mimic.R
\name{import_mimic}
\alias{import_mimic}
\title{Import and Format MIMIC-IV}
\usage{
import_mimic(path, verbose = interactive())
}
\arguments{
\item{path}{\code{character(1)}, path to the root MIMIC-IV folder (that contains
the subfolders: core and hosp).}

\item{verbose}{\code{logical(1)}, if \code{TRUE} progress messages are shown.}
}
\value{
\code{data.frame}, same as \code{x} but with an added column \code{Excluded} and
3 attributes \code{"message"}, \code{"n_cases"}, \code{"n_cbc"} that describe the
processing and the kept cases and complete blood counts (CBC)
}
\description{
Import and format MIMIC-IV dataset as done in the AMPEL project.
}
\details{
For the MIMIC-IV dataset the files \code{core/patients.csv}, \code{core/transfers.csv},
\code{hosp/labevents.csv}, and \code{hosp/diagnoses_icd.csv} are needed.
}
\examples{
\dontrun{
mimic <- import_mimic("path/to/mimic-iv-1.0")
}
}
\references{
\itemize{
\item Johnson, A., Bulgarelli, L., Pollard, T., Horng, S., Celi, L. A., & Mark, R.
(2021). MIMIC-IV (version 1.0). PhysioNet. \doi{10.13026/s6n6-xd98}.
\item Documentation \url{https://mimic.mit.edu/docs/iv/}
}
}
\author{
Daniel Steinbach \href{mailto:daniel.steinbach@medizin.uni-leipzig.de}{daniel.steinbach@medizin.uni-leipzig.de} and
Sebastian Gibb \href{mailto:mail@sebastiangibb.de}{mail@sebastiangibb.de}
}
