% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labcodes.R
\docType{data}
\name{labcodes}
\alias{labcodes}
\title{Laboratory Codes}
\format{
A \code{data.frame} with 3 columns/variables:
\describe{
\item{Code}{\code{character}, code/column name in \code{\link{sbcdata}}.}
\item{Description}{\code{character}, description.}
\item{Unit}{\code{character}, unit.}
}
}
\usage{
labcodes
}
\description{
This dataset contains details about the different laboratory measurements.
}
\keyword{datasets}
