% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exclude.R
\name{sbc_exclude_entries}
\alias{sbc_exclude_entries}
\title{AMPEL Exclusion}
\usage{
sbc_exclude_entries(x, time = c(-Inf, Inf) * 3600)
}
\arguments{
\item{x}{\code{data.table}, in the format described in \code{\link{sbcdata}}}

\item{time}{\code{numeric(2)}, keep just entries between \code{time[1]} and
\code{time[2]}, in seconds.}
}
\value{
\code{data.table}, same as \code{x} but with an added column \code{Excluded} and
3 attributes, namely \code{"exclude_message"}, \code{"exclude_cases"}, \code{"exclude_cbc"}
that describe the processing and the kept cases and
complete blood counts (CBC)
}
\description{
Run the exclusion as done in the AMPEL project.
}
\examples{
x <- sbc_exclude_entries(sbcdata)
attributes(x)
}
\author{
Sebastian Gibb \href{mailto:mail@sebastiangibb.de}{mail@sebastiangibb.de}
}
