% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label.R
\name{sbc_label}
\alias{sbc_label}
\title{Label BCs}
\usage{
sbc_label(x, time = 6 * 3600)
}
\arguments{
\item{x}{\code{data.table}, in the format described in \code{\link{sbcdata}}}

\item{time}{\code{numeric(1)}, label entries with \code{"SecToIcu"}
below this time and \code{Diagnosis == "Sepsis"} as "Sepsis" or as "Control"
otherwise.}
}
\value{
\code{data.table}, same as \code{x} but with an added column \code{Label} and
3 attributes, namely \code{"label_message"}, \code{"label_cases"}, \code{"label_cbc"} that
describe the labeling and the affected cases and complete blood counts (CBC).
}
\description{
Run the labeling as done in the AMPEL project.
}
\examples{
x <- sbc_exclude_entries(sbcdata)
attributes(x)
x <- sbc_label(x)
attributes(x)
}
\author{
Sebastian Gibb \href{mailto:mail@sebastiangibb.de}{mail@sebastiangibb.de}
}
