% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{sbc_preprocess}
\alias{sbc_preprocess}
\title{AMPEL Preprocessing}
\usage{
sbc_preprocess(x)
}
\arguments{
\item{x}{\code{data.table}, in the format described in \code{\link{sbcdata}}}
}
\value{
\code{data.table}, same as \code{x} but with additional columns \code{Excluded} and
\code{Label}. Also 6 attributes are added:
\code{"exclude_message"}, \code{"exclude_cases"}, \code{"exclude_cbc"} that
describe the exclusion and the kept cases and complete blood counts (CBC).
\code{"label_message"}, \code{"label_cases"}, \code{"label_cbc"} that
describe the labeling and the affected cases and complete blood counts (CBC).
}
\description{
Run the preprocessing as done in the AMPEL project.
}
\author{
Sebastian Gibb \href{mailto:mail@sebastiangibb.de}{mail@sebastiangibb.de}
}
