% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbcdata-package.R
\docType{package}
\name{sbcdata-package}
\alias{sbcdata-package}
\title{sbcdata: Laboratory Diagnostics from Septic and Non-septic Patients Used in the AMPEL Project}
\description{
The data were collected as part of the AMPEL project (Analysis and Reporting System for the Improvement of Patient Safety through Real-Time Integration of Laboratory Findings, \url{https://ampel.care}). They include administration data, sepsis labels based on ICD10 codes and laboratory diagnostics from patients admitted to the University Hospital Leipzig between Januar 2014 and December 2021 and to the University Hospital Greifswald between Januar 2015 and December 2020, respectively. Additionally it provides functions to generate a similar dataset from MIMIC-IV (Johnson et al. 2021 \doi{10.13026/S6N6-XD98}).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ampel-leipzig/sbcdata}
  \item Report bugs at \url{https://github.com/ampel-leipzig/sbcdata/issues/}
}

}
\author{
\strong{Maintainer}: Sebastian Gibb \email{mail@sebastiangibb.de} (\href{https://orcid.org/0000-0001-7406-4443}{ORCID})

Authors:
\itemize{
  \item Paul Ahrens \email{paul.ahrens@medizin.uni-leipzig.de} [data contributor]
  \item Daniel Steinbach \email{daniel.steinbach@medizin.uni-leipzig.de} [data contributor]
  \item Maria Schmidt \email{maria.schmidt@medizin.uni-leipzig.de} [data contributor]
  \item Thorsten Kaiser \email{thorsten.kaiser@medizin.uni-leipzig.de} [data contributor]
}

Other contributors:
\itemize{
  \item Mark Wernsdorfer \email{mark.wernsdorfer@medizin.uni-leipzig.de} [data contributor]
  \item Matthias Nauck \email{matthias.nauck@med.uni-greifswald.de} [data contributor]
  \item Stefan Bollmann \email{stefan.bollmann@med.uni-greifswald.de} [data contributor]
  \item Thomas Hildebrandt \email{thomas.hildebrandt@med.uni-greifswald.de} [data contributor]
}

}
\keyword{internal}
