% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbcdata.R
\docType{data}
\name{sbcdata}
\alias{sbcdata}
\title{Laboratory Diagnostics of Patients at University Hospitals}
\format{
A \code{data.table} with 18 columns/variables:
\describe{
\item{Id}{\code{integer}, identification number of the case, unique for each
center.}
\item{Age}{\code{integer}, age of the patient in years.}
\item{Sex}{\code{character}, "\code{W}" for female and \code{"M"} for male.}
\item{Diagnosis}{\code{character}, diagnosis, could be \code{"Control"}, \code{"SIRS"}
and \code{"Sepsis"}. See below for details.}
\item{Center}{\code{character}, center, one of \code{"Greifswald"} or \code{"Leipzig"}.}
\item{Set}{\code{character}, center, one of \code{"Training"} or \code{"Validation"}.}
\item{Sender}{\code{character}, sender/origin which send the blood sample to the
laboratory. See \code{\link{sendercodes}} for a description of all possible codes.}
\item{Epsiode}{\code{integer}, counter for episodes on intensive care units,
is incremented by one after each discharge from the intensive care unit.}
\item{Time}{\code{integer}, (relative) time when the blood was analysed.
The first timepoint for each case is always set to zero.}
\item{TragetIcu}{\code{character}, the name/type of the intensive care unit where
the patient/case has to be admitted to. See \code{\link{sendercodes}} for a
description of all possible codes.}
\item{SecToIcu}{\code{integer}, time in seconds until the patient/case has to be
admitted to the \code{TargetIcu} intensive care unit.
This time is negative if the patient/case is already on the
intensive care unit \code{TargetIcu}.}
\item{CRP}{\code{double}, C-reactive protein in mg/l.}
\item{HGB}{\code{double}, hemoglobin in mmol/l.}
\item{MCV}{\code{double}, mean corpuscular volume in fl.}
\item{PLT}{\code{double}, procalcitonin in Gpt/l.}
\item{RBC}{\code{double}, platelets in Tpt/l.}
\item{WBC}{\code{double}, red blood count in Gpt/l.}
\item{PCT}{\code{double}, white blood count in ng/ml.}
}
}
\source{
{
\subsection{Data University Hospital Leipzig}{

\describe{
\item{R package}{Sebastian Gibb}
\item{Data Processing}{Maria Schmidt, Paul Ahrens and Mark Wernsdorfer}
\item{Laboratory Data Collection/Extraction}{Maria Schmidt, Thorsten Kaiser}
\item{Administration Data Extraction}{Maria Schmidt, Thorsten Kaiser}
\item{Reference Ethic Committee}{214/18ek}
}}
}

{
\subsection{Data University Medicine Greifswald}{

\describe{
\item{Data Processing/R package}{Sebastian Gibb}
\item{Laboratory Data Collection/Extraction}{Matthias Nauck and Stefan Bollmann}
\item{Administration Data Extraction}{Thomas Hildebrandt}
\item{Reference Ethic Committee}{BB133/10}
}}
}
}
\usage{
sbcdata
}
\description{
This dataset includes laboratory diagnostics for the
complete blood counts without differentiation,
C-reactive protein and procalictonin for
patients admitted at the University Hospital Leipzig from
2014 to 2019 as Training and from 2020 to 2021 as Validation set,
respectively.
For external validation the same laboratory values were taken
from 2015 to 2020 at the University Hospital Greifswald.
}
\details{
The \code{Diagnosis} was based on ICD10 codes. \code{"Sepsis"} was assumed for:
\itemize{
\item{A02.1}
\item{A20.7}
\item{A22.7}
\item{A24.1}
\item{A26.7}
\item{A32.7}
\item{A39.2, A39.3, A39.4}
\item{A40.0, A40.1, A40.2, A40.3, A40.8, A40.9}
\item{A41.0, A41.1, A41.2, A41.3, A41.4, A41.51, A41.52, A41.58, A41.8,
A41.9}
\item{A42.7}
\item{B37.7}
\item{R57.2}
}
If the ICD10 code was R65.x without any of the sepsis-related codes above the
\code{Diagnoses} \code{"SIRS"} was used. Everything else is labeled as "\code{Control}".

For the \code{Center} \code{"Greifswald"} there are a few entries with duplicated
time points \code{Time} for the same \code{Id} and \code{Sender} with different laboratory
values. This happens due to the analyses of multiple blood samples from the
same patient at the same time in the same run of the analyser.
It could not be decided which one is the correct/better one so removal is
suggested. An example could be found below.

In the \code{Center} \code{"Greifswald"} the sender codes are not as detailed as in
\code{"Leipzig"}. That's why \code{"OTHER"} occures more often and \code{"AMB"} fewer
than in \code{"Leipzig"}.

At the \code{Center} \code{"Greifswald"} the admission/discharge timepoint was recorded
in and extracted from the clinical information system. These data were not
available for the \code{Center} \code{"Leipzig"}. There the first/last blood sample
taken on an intensive care unit was taken as timepoint for
admission/discharge (which is not necessarly part of the dataset).
That's why the first blood sample on an intensive care
unit could have a \code{SecToIcu} of zero in contrast to a negative value for
\code{Center} \code{"Greifswald"}. If needed this could be harmonized by adding the
first \code{SecToIcu} for an intensive care unit to all \code{SecToIcu} for each
\code{Episode} in the \code{"Greifswald"} and/or \code{"Leipzig"} subset.
An example could be found below.

In a few cases there is a mismatch between the timepoint of
admission/discharge extracted from the clinical information system at
\code{Center} \code{"Greifswald"} and the entry in \code{Center}. It could happen that the
\code{Sender} is a non-ICU ward and the \code{SecToIcu} time is negative. According to
the admission data the patient was already on an ICU but the laboratory order
was taken from a non-ICU ward. Mostly this time difference is around a few
minutes (a blood sample was taken on the non-ICU ward, but the analysis in
the laboratory started after transfer to the ICU).
}
\examples{
## remove duplicate laboratory entries (see text above for explanation)
greifswald <- subset(sbcdata, Center == "Greifswald")
dup <- duplicated(greifswald[, .(Id, Time)]) |
    duplicated(greifswald[, .(Id, Time)], fromLast = TRUE)
mean(dup)
greifswald <- greifswald[!dup,]

## adjust SecToIcu for subset Greifswald (see text above for explanation)
greifswald <- subset(sbcdata, Center == "Greifswald")
## create helper columns
greifswald[, isNewWard := (
    c(FALSE, Id[-1] == Id[-.N]) &         # same case
    c(FALSE, Sender[-1] != Sender[-.N])   # new ward
)]
greifswald[, isIcuAdmission := isNewWard & grepl("ICU", Sender)]
## recalculate SecToIcu
greifswald[, SecToIcu := SecToIcu - SecToIcu[isIcuAdmission][Episode]]
## drop helper columns
greifswald[, `:=` (isNewWard = NULL, isIcuAdmission = NULL)]
}
\keyword{datasets}
