% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sendercodes.R
\docType{data}
\name{sendercodes}
\alias{sendercodes}
\title{Sender Codes}
\format{
A \code{data.frame} with 2 columns/variables:
\describe{
\item{Sender}{\code{character}, sender code}
\item{Description}{\code{character}, description.}
}
}
\usage{
sendercodes
}
\description{
This dataset describes sender of blood samples, e.g. different types
of wards/units.
}
\keyword{datasets}
