% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exclude.R
\name{exclude_entries}
\alias{exclude_entries}
\title{AMPEL Preprocessing/Exclusion}
\usage{
exclude_entries(x, time = c(6, 72) * 3600)
}
\arguments{
\item{x}{\code{data.table}, in the format described in \code{\link{sbcdata}}}

\item{time}{\code{numeric(2)}, keep just entries between \code{time[1]} and
\code{time[2]}, in seconds.}
}
\value{
\code{data.frame}, same as \code{x} but with an added column \code{Excluded} and
3 attributes \code{"message"}, \code{"n_cases"}, \code{"n_cbc"} that describe the
processing and the kept cases and complete blood counts (CBC)
}
\description{
Run the preproccesing/exclusion as done in the AMPEL project.
}
\examples{
x <- exclude_entries(sbcdata)
attributes(x)
}
\author{
Sebastian Gibb \href{mailto:mail@sebastiangibb.de}{mail@sebastiangibb.de}
}
